+++ src/SyncAPIConnector/codes/REQUEST_STATUS.cs
@@ -4,17 +4,17 @@ namespace xAPI.Codes;
 
 public class REQUEST_STATUS : BaseCode
 {
-    public const long ERROR_CODE = 0;
-    public const long PENDING_CODE = 1;
-    public const long ACCEPTED_CODE = 3;
-    public const long REJECTED_CODE = 4;
+    public const int ERROR_CODE = 0;
+    public const int PENDING_CODE = 1;
+    public const int ACCEPTED_CODE = 3;
+    public const int REJECTED_CODE = 4;
 
     public static readonly REQUEST_STATUS ERROR = new(ERROR_CODE);
     public static readonly REQUEST_STATUS PENDING = new(PENDING_CODE);
     public static readonly REQUEST_STATUS ACCEPTED = new(ACCEPTED_CODE);
     public static readonly REQUEST_STATUS REJECTED = new(REJECTED_CODE);
 
-    public REQUEST_STATUS(long code)
+    public REQUEST_STATUS(int code)
         : base(code)
     {
     }
